Update serivce principal instructions to api://
There was an update to Azure CLI that requires us to use `api://` instead of `https://`. Until we can update the VM, we are going to update the instructions to use `api://`. 

Additionally, at the moment there seems to an issue with certificate vertification with Azure CLI, I also added a note that suggests waiting a few minutes and trying again. 

Related to #90 and #86