Use Object ID to reference a service principal
In Azure CLI version 2.29 and later, running the command "az keyvault set-policy"
with the parameter --spn set to 'api:<spName>', where <spName> is the name of
the service principal, fails. It results in the error message "Unable to find user with
spn ‘api:<spName>'. Unable to get object id from principal name."

This is due to the fact that the ‘name’ property of the service principal is deprecated.
Setting --spn to the app ID works. However, this may not work with older versions
of Azure CLI, including version 2.17.1, which is used in the hosted lab VM.

To make sure this lab’s instructions are valid for both the hosted lab’s VM and the
current version of Azure CLI, the instructions are updated to:

- show how to find the object ID of the service principal using Azure CLI, and

- reference the service principal by its object ID, instead of its name in the
"az keyvault set-policy" command.